+++ Exercise 1/ProblemInstance.java
@@ -1,9 +1,9 @@
 
 public class ProblemInstance {
 	private int numJobs;
-	private int[][] jobs; //size = [num_jobs][2], for every job [0] is the length, [1] is the due time
+	private double[][] jobs; //size = [num_jobs][2], for every job [0] is the length, [1] is the due time
 	
-	public ProblemInstance(int numJobs, int[][] jobs) {
+	public ProblemInstance(int numJobs, double[][] jobs) {
 		this.numJobs = numJobs;
 		this.jobs = jobs;
 	}
@@ -12,7 +12,7 @@ public class ProblemInstance {
 		return numJobs;
 	}
 	
-	public int[][] getJobs() {
+	public double[][] getJobs() {
 		return jobs;
 	}
 }
